Reorganize scenario code
