+++ redskyctl/internal/commands/generate/experiment/generator.go
@@ -22,13 +22,10 @@ import (
 	redskyappsv1alpha1 "github.com/redskyops/redskyops-controller/api/apps/v1alpha1"
 	redskyv1beta1 "github.com/redskyops/redskyops-controller/api/v1beta1"
 	meta2 "github.com/redskyops/redskyops-controller/internal/meta"
-	"github.com/redskyops/redskyops-controller/internal/setup"
 	"github.com/redskyops/redskyops-controller/pkg/application"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/kustomize/api/filesys"
 )
 
@@ -130,22 +127,6 @@ func (g *Generator) Generate() (*corev1.List, error) {
 	return list, nil
 }
 
-// findOrAddExperiment returns the experiment from the supplied list, creating it if it does not exist.
-func findOrAddExperiment(list *corev1.List) *redskyv1beta1.Experiment {
-	var exp *redskyv1beta1.Experiment
-	for i := range list.Items {
-		if p, ok := list.Items[i].Object.(*redskyv1beta1.Experiment); ok {
-			exp = p
-			break
-		}
-	}
-	if exp == nil {
-		exp = &redskyv1beta1.Experiment{}
-		list.Items = append(list.Items, runtime.RawExtension{Object: exp})
-	}
-	return exp
-}
-
 // labelExperiment adds application labels to the experiment
 func labelExperiment(app *redskyappsv1alpha1.Application, exp *redskyv1beta1.Experiment) {
 	// This function adds labels to the experiment, the trial template and (if present)
@@ -166,101 +147,4 @@ func labelExperiment(app *redskyappsv1alpha1.Application, exp *redskyv1beta1.Exp
 	if len(app.Scenarios) == 1 {
 		addExpLabel(redskyappsv1alpha1.LabelScenario, app.Scenarios[0].Name)
 	}
-}
-
-// ensureSetupServiceAccount ensures that we are using an explicit service account for setup tasks.
-func ensureSetupServiceAccount(list *corev1.List) {
-	// Return if we see an explicit service account name
-	exp := findOrAddExperiment(list)
-	saName := &exp.Spec.TrialTemplate.Spec.SetupServiceAccountName
-	if *saName != "" {
-		return
-	}
-	*saName = "redsky-setup"
-
-	// Add the actual service account to the list
-	list.Items = append(list.Items,
-		runtime.RawExtension{Object: &corev1.ServiceAccount{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: *saName,
-			},
-		}},
-	)
-}
-
-// ensurePrometheus adds Prometheus configuration to the supplied list.
-func ensurePrometheus(list *corev1.List) {
-	// Return if we see the Prometheus setup task
-	exp := findOrAddExperiment(list)
-	trialSpec := &exp.Spec.TrialTemplate.Spec
-	for _, st := range trialSpec.SetupTasks {
-		if setup.IsPrometheusSetupTask(&st) {
-			return
-		}
-	}
-
-	// Add the missing setup task
-	trialSpec.SetupTasks = append(trialSpec.SetupTasks, redskyv1beta1.SetupTask{
-		Name: "monitoring",
-		Args: []string{"prometheus", "$(MODE)"},
-	})
-
-	// Ensure there is an explicit service account configured
-	ensureSetupServiceAccount(list)
-
-	// Append the cluster role and binding for Prometheus
-	list.Items = append(list.Items,
-		runtime.RawExtension{Object: &rbacv1.ClusterRole{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "redsky-prometheus",
-			},
-			Rules: []rbacv1.PolicyRule{
-				// Required to manage the Prometheus resources in the setup task
-				{
-					Verbs:     []string{"get", "create", "delete"},
-					APIGroups: []string{"rbac.authorization.k8s.io"},
-					Resources: []string{"clusterroles", "clusterrolebindings"},
-				},
-				{
-					Verbs:     []string{"get", "create", "delete"},
-					APIGroups: []string{""},
-					Resources: []string{"serviceaccounts", "services", "configmaps"},
-				},
-				{
-					Verbs:     []string{"get", "create", "delete", "list", "watch"},
-					APIGroups: []string{"apps"},
-					Resources: []string{"deployments"},
-				},
-
-				// Permissions we need to delegate to Prometheus runtime (prometheus-server-rbac.yaml)
-				{
-					Verbs:     []string{"list", "watch", "get"},
-					APIGroups: []string{""},
-					Resources: []string{"nodes", "nodes/metrics", "nodes/proxy", "services"},
-				},
-				{
-					Verbs:     []string{"list", "watch"},
-					APIGroups: []string{""},
-					Resources: []string{"pods"},
-				},
-			},
-		}},
-		runtime.RawExtension{Object: &rbacv1.ClusterRoleBinding{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "redsky-setup-prometheus",
-			},
-			RoleRef: rbacv1.RoleRef{
-				APIGroup: "rbac.authorization.k8s.io",
-				Kind:     "ClusterRole",
-				Name:     "redsky-prometheus",
-			},
-			Subjects: []rbacv1.Subject{
-				{
-					Kind: "ServiceAccount",
-					Name: exp.Spec.TrialTemplate.Spec.SetupServiceAccountName,
-				},
-			},
-		}},
-	)
-
 }
